rc: build new instance from ground up, following void's scheme closely
This is meant to be an experiment (testing on a f2fs system to see
what's wrong with the old Arch initscript). However, this may be merged
into our mainline. If that *does* happen, I'll merge some of the old
scirpts (e.g. tmpfiles-setup, lvm) back, but some sv.d scripts (like netfs)
will need to be removed permanently.